\name{persomind}
\alias{persomind}
\title{persomind}
\author{Dominique Makowski}
\description{
A dataset containing personality and mindfulness trait data from an online questionnaire.

If you use this dataset for a publication, please refer to it as:
"persomind-1.0".
}
\examples{
library(neuropsychology)

df <- persomind
}
