\name{personality}
\alias{personality}
\title{personality}
\author{Dominique Makowski}
\description{
A dataset containing normal and pathological personality traits data from an online questionnaire.

If you use this dataset for a publication, please refer to it as:
"personality-1.0".
}
\examples{
library(neuropsychology)

df <- personality
}
